+++ api/apps.py
@@ -1,0 +1,8 @@
+"""API apps."""
+from django.apps import AppConfig
+
+
+class ApiConfig(AppConfig):
+    """Configuration for the API."""
+
+    name = 'api'
