+++ mission_control/tests/test_urls.py
@@ -53,30 +53,6 @@ class TestMissionControlURLs(TestCase):
             resolve('/mission-control/rover-list/').view_name,
             'mission-control:rover_list')
 
-    def test_rovers_reverse(self):
-        """mission-control:rover-list should reverse to /mission-control/rovers/.""" # noqa
-        self.assertEqual(
-            reverse('mission-control:rover-list'),
-            '/mission-control/rovers/')
-
-    def test_rovers_resolve(self):
-        """/mission-control/rovers/ should resolve to mission-control:rover-list.""" # noqa
-        self.assertEqual(
-            resolve('/mission-control/rovers/').view_name,
-            'mission-control:rover-list')
-
-    def test_block_diagrams_reverse(self):
-        """mission-control:blockdiagram-list should reverse to /mission-control/block-diagrams/.""" # noqa
-        self.assertEqual(
-            reverse('mission-control:blockdiagram-list'),
-            '/mission-control/block-diagrams/')
-
-    def test_block_diagrams_resolve(self):
-        """/mission-control/block-diagrams/ should resolve to mission-control:blockdiagram-list.""" # noqa
-        self.assertEqual(
-            resolve('/mission-control/block-diagrams/').view_name,
-            'mission-control:blockdiagram-list')
-
     def test_rover_settings_reverse(self):
         """'mission-control:rover_settings' should reverse to /mission-control/rover-settings/.""" # noqa
         self.assertEqual(
