+++ mission_control/views.py
@@ -2,11 +2,9 @@
 from django.contrib.auth.decorators import login_required
 from django.core.urlresolvers import reverse
 from django.shortcuts import render, redirect
-from django_filters.rest_framework import DjangoFilterBackend
 from .models import Rover, BlockDiagram
-from rest_framework import viewsets, permissions, serializers
 from rest_framework.renderers import JSONRenderer
-from .serializers import RoverSerializer, BlockDiagramSerializer
+from .serializers import BlockDiagramSerializer
 from django.views.decorators.csrf import ensure_csrf_cookie
 from django.shortcuts import get_object_or_404
 from .forms import RoverForm
@@ -58,43 +56,3 @@ def rover_settings(request, pk):
         'name': rover.name,
         'form': form
     })
-
-
-class RoverViewSet(viewsets.ModelViewSet):
-    """API endpoint that allows rovers to be viewed or edited."""
-
-    serializer_class = RoverSerializer
-    permission_classes = (permissions.IsAuthenticated, )
-    filter_backends = (DjangoFilterBackend,)
-    filter_fields = ('name',)
-
-    def get_queryset(self):
-        """The list of rovers for the user."""
-        return Rover.objects.filter(owner=self.request.user.id)
-
-    def perform_create(self, serializer):
-        """Perform the create operation."""
-        user = self.request.user
-        serializer.save(owner=user)
-
-
-class BlockDiagramViewSet(viewsets.ModelViewSet):
-    """API endpoint that allows block diagrams to be viewed or edited."""
-
-    queryset = BlockDiagram.objects.all()
-    serializer_class = BlockDiagramSerializer
-    permission_classes = (permissions.IsAuthenticated, )
-    filter_backends = (DjangoFilterBackend,)
-    filter_fields = ('user', 'name')
-
-    def perform_create(self, serializer):
-        """Perform the create operation."""
-        user = self.request.user
-        serializer.save(user=user)
-
-    def perform_update(self, serializer):
-        """Perform the update operation."""
-        if self.get_object().user.id is not self.request.user.id:
-            raise serializers.ValidationError(
-                'You may only modify your own block diagrams')
-        serializer.save()
